+++ Cpp-C/Eta/Applications/Examples/VAProvider/rsslItemHandler.c
@@ -483,7 +483,7 @@ static RsslRet closeItemStream(RsslReactor *pReactor, RsslReactorChannel *pReact
 	/* remove original item request information */
 	if (itemRequestInfo)
 	{
-		printf("Closing item stream id %d with item name: %.*s\n", itemRequestInfo->StreamId, strlen(itemRequestInfo->Itemname), itemRequestInfo->Itemname);
+		printf("Closing item stream id %d with item name: %.*s\n", itemRequestInfo->StreamId, (int)strlen(itemRequestInfo->Itemname), itemRequestInfo->Itemname);
 		if (freeItemReqInfo(pReactor, itemRequestInfo) != RSSL_RET_SUCCESS)
 			return RSSL_RET_FAILURE;
 	}
@@ -513,7 +513,7 @@ void closeItemChnlStreams(RsslReactor *pReactor, RsslReactorChannel* pReactorCha
 		{
 			itemRequestInfo = &itemRequestInfoList[i];
 			/* remove original item request information */
-			printf("Closing item stream id %d with item name: %.*s and domainType: %u\n", itemRequestInfo->StreamId, strlen(itemRequestInfo->Itemname), itemRequestInfo->Itemname, itemRequestInfo->domainType);
+			printf("Closing item stream id %d with item name: %.*s and domainType: %u\n", itemRequestInfo->StreamId, (int)strlen(itemRequestInfo->Itemname), itemRequestInfo->Itemname, itemRequestInfo->domainType);
 			freeItemReqInfo(pReactor, itemRequestInfo);
 		}
 	}
@@ -825,11 +825,11 @@ RsslRet processSingleItemRequest(RsslReactor *pReactor, RsslReactorChannel* pRea
 
 	if (msg->requestMsg.flags & RSSL_RQMF_PRIVATE_STREAM)
 	{
-		printf("\nReceived Private Stream Item Request for %.*s (streamId=%d) on domain %s\n", strlen(itemReqInfo->Itemname), itemReqInfo->Itemname, msg->msgBase.streamId,  rsslDomainTypeToString(itemReqInfo->domainType));
+		printf("\nReceived Private Stream Item Request for %.*s (streamId=%d) on domain %s\n", (int)strlen(itemReqInfo->Itemname), itemReqInfo->Itemname, msg->msgBase.streamId,  rsslDomainTypeToString(itemReqInfo->domainType));
 	}
 	else
 	{
-	printf("\nReceived Item Request for %.*s (streamId=%d) on domain %s\n", strlen(itemReqInfo->Itemname), itemReqInfo->Itemname, msg->msgBase.streamId,  rsslDomainTypeToString(itemReqInfo->domainType));
+	printf("\nReceived Item Request for %.*s (streamId=%d) on domain %s\n", (int)strlen(itemReqInfo->Itemname), itemReqInfo->Itemname, msg->msgBase.streamId,  rsslDomainTypeToString(itemReqInfo->domainType));
 	}
 
 	/* send item response */
@@ -973,11 +973,11 @@ RsslRet processBatchRequest(RsslReactor *pReactor, RsslReactorChannel* pReactorC
 
 				if (msg->requestMsg.flags & RSSL_RQMF_PRIVATE_STREAM)
 				{
-					printf("\nReceived Private Stream Item Request for %.*s (streamId=%d) on domain %s\n", strlen(itemReqInfo->Itemname), itemReqInfo->Itemname, itemStream,  rsslDomainTypeToString(itemReqInfo->domainType));
+					printf("\nReceived Private Stream Item Request for %.*s (streamId=%d) on domain %s\n", (int)strlen(itemReqInfo->Itemname), itemReqInfo->Itemname, itemStream,  rsslDomainTypeToString(itemReqInfo->domainType));
 				}
 				else
 				{
-					printf("\nReceived Item Request for %.*s (streamId=%d) on domain %s\n", strlen(itemReqInfo->Itemname), itemReqInfo->Itemname, itemStream,  rsslDomainTypeToString(itemReqInfo->domainType));
+					printf("\nReceived Item Request for %.*s (streamId=%d) on domain %s\n", (int)strlen(itemReqInfo->Itemname), itemReqInfo->Itemname, itemStream,  rsslDomainTypeToString(itemReqInfo->domainType));
 				}
 
 				/* send item response/refresh if required */
